backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
openpyxl==3.1.5
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.4
pandas==2.2.3
numpy==2.2.1
